Move `extern "C"` block to remacs-sys
In a effort to refactor lisp.rs i moved some extern functions and
statics to remacs-sys.
