+++ rust_src/src/strings.rs
@@ -2,9 +2,9 @@ use std::ptr;
 
 use libc;
 
-use lisp::{LispObject, Qnil, SBYTES, SSDATA, STRING_MULTIBYTE, STRINGP, CHECK_STRING};
+use lisp::{LispObject, Qnil, SBYTES, STRINGP, CHECK_STRING};
 use lists::NILP;
-use remacs_sys::Lisp_Object;
+use remacs_sys::{Lisp_Object, SSDATA, STRING_MULTIBYTE};
 
 extern "C" {
     fn make_string(s: *const libc::c_char, length: libc::ptrdiff_t) -> Lisp_Object;
